requirements-tg-bot.txt
# for the tg-bot
python-telegram-bot==21.3
